refactor: rename operations page
