+++ bciers/apps/administration/app/idir/cas_analyst/operations/page.tsx
@@ -1,8 +1,7 @@
 // 🚩 flagging that for shared routes between roles, "Page" code is a component for code maintainability
 import { OperationsSearchParams } from "@/app/components/operations/types";
-import OperationsPage, {
-  InternalUserOperationsLayout,
-} from "@/administration/app/components/operations/Operations";
+import OperationDataGridPage from "@/administration/app/components/operations/OperationDataGridPage";
+import { InternalUserOperationDataGridLayout } from "@/administration/app/components/operations/OperationLayouts";
 
 export default async function Page({
   searchParams,
@@ -10,8 +9,8 @@ export default async function Page({
   searchParams: OperationsSearchParams;
 }) {
   return (
-    <InternalUserOperationsLayout>
-      <OperationsPage searchParams={searchParams} />
-    </InternalUserOperationsLayout>
+    <InternalUserOperationDataGridLayout>
+      <OperationDataGridPage searchParams={searchParams} />
+    </InternalUserOperationDataGridLayout>
   );
 }
