+++ bciers/apps/administration/tests/components/operations/OperationDataGridPage.test.tsx
@@ -4,7 +4,7 @@ import {
   useRouter,
   useSearchParams,
 } from "@bciers/testConfig/mocks";
-import Operations from "apps/administration/app/components/operations/Operations";
+import Operations from "apps/administration/app/components/operations/OperationDataGridPage";
 import { FrontEndRoles } from "@bciers/utils/enums";
 
 useRouter.mockReturnValue({
